requirements.txt
# Framework Web
fastapi
uvicorn[standard]
pydantic-settings
httpx

# LangChain
langchain
langchain-community
langchain-core
langchain_ollama
langchain-google-genai

# Vector Store
chromadb

# Document Processing
pandas
openpyxl
pymupdf
